Added test editor task.
