+++ editor/src/core/editor.h
@@ -3,6 +3,8 @@
 #include <SDL2/SDL.h>
 #include <glad/glad.h>
 
+#include "SquidTasks/TaskManager.h"
+
 class Editor {
   public:
     bool Initialize();
@@ -10,9 +12,13 @@ class Editor {
     bool IsRunning() const;
     void Shutdown();
 
+    static float GetCurrentTime();
+
   private:
     int windowWidth = 800;
     int windowHeight = 600;
+
+    TaskManager mainTasks;
 
     bool InitializeSDL();
     bool InitializeImGui();
