+++ editor/src/main.cpp
@@ -1,20 +1,4 @@
 #include "core/editor.h"
-#include "../engine/src/core/utils/logger.h"
-
-#include "SquidTasks/Task.h"
-
-using namespace Squid;
-
-Task<> TestTask() {
-    TASK_NAME(__FUNCTION__ );
-
-    int number = 0;
-
-    while (true) {
-//        rbe_logger_info("number = %d", number++);
-        co_await Suspend();
-    }
-}
 
 int main(int argv, char** args) {
     Editor editor;
@@ -22,11 +6,8 @@ int main(int argv, char** args) {
         return -1;
     }
 
-    auto TestTaskHandle = TestTask();
-
     while (editor.IsRunning()) {
         editor.Update();
-        TestTaskHandle.Resume();
     }
 
     editor.Shutdown();
